pyproject.toml
[tool.poetry]
name = "fragalysis-ispyb-target-access-authenticator"
version = "0.1.0"
package-mode = false
description = "The Fragalysis Stack Python/FastAPI ISPyB Target Access Authenticator"
authors = ["Alan Christie <achristie@einformaticsmatters.com>"]
readme = "README.md"

[tool.poetry.dependencies]
# Locked versions (for now)
fastapi = "0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
# We're stuck on ISPyB 4.x to avoid the error: -
#   ImportError: cannot import name 'ISPyBNoResultException' from 'ispyb.connector.mysqlsp.main'
ispyb = "4.13.2"
# ispyb (4.13.2) depends on mysql-connector (<2.2.3)
mysql-connector = "2.1.7"
# Flexible versions
prometheus-client = "^0.22.1"
pymemcache = "^4.0.0"
pydantic = "^2.10.6"
pymysql = "^1.1.1"
python = "^3.12"
python-dateutil = "^2.9.0"
sshtunnel = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pylint = "^3.3.7"
httpie = "^3.2.4"
isort = "^6.0.1"
black = "^25.1.0"

[tool.pylint.'MESSAGES CONTROL']
disable = "too-few-public-methods,too-many-arguments,too-many-positional-arguments,missing-function-docstring,missing-module-docstring,missing-class-docstring"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
